Update directory structure; fix #51
