+++ src/core/core_structures.jl
